Move Conditional.*, LayEval.*, fmt_config.*, fmt_utility.*, printer.* into formatting
