+++ src/Battery.hpp
@@ -6,8 +6,8 @@
 # include <vector>
 # include <optional>
 
-# include "LazyEval.hpp"
-# include "Conditional.hpp"
+# include "formatting/LazyEval.hpp"
+# include "formatting/Conditional.hpp"
 
 # include "dep/fmt/include/fmt/format.h"
