+++ src/Makefile
@@ -89,16 +89,19 @@ all_builds: release_build release_no_py_build
 debug_build:
 	mkdir -p $@
 	mkdir -p $@/Callback
+	mkdir -p $@/formatting
 	$(MAKE) BUILD_DIR=$@ DEBUG=true  PYTHON=true
 
 release_build:
 	mkdir -p $@
 	mkdir -p $@/Callback
+	mkdir -p $@/formatting
 	$(MAKE) BUILD_DIR=$@ DEBUG=false PYTHON=true
 
 release_no_py_build:
 	mkdir -p $@
 	mkdir -p $@/Callback
+	mkdir -p $@/formatting
 	$(MAKE) BUILD_DIR=$@ DEBUG=false PYTHON=false
 
 install: $(BUILD_DIR)/swaystatus
