+++ src/formatting/fmt_utility.cc
