+++ src/print_battery.cc
@@ -22,7 +22,7 @@
 #include "utility.h"
 #include "process_configuration.h"
 #include "handle_click_events.h"
-#include "printer.hpp"
+#include "formatting/printer.hpp"
 #include "Battery.hpp"
 #include "print_battery.h"
