+++ src/print_load.cc
@@ -12,7 +12,7 @@
 #include "utility.h"
 #include "process_configuration.h"
 #include "handle_click_events.h"
-#include "printer.hpp"
+#include "formatting/printer.hpp"
 #include "print_load.h"
 
 using swaystatus::print;
