+++ src/print_volume.cc
@@ -4,7 +4,7 @@
 #include "alsa.h"
 #include "process_configuration.h"
 #include "handle_click_events.h"
-#include "printer.hpp"
+#include "formatting/printer.hpp"
 #include "print_volume.h"
 
 using swaystatus::print;
